+++ docs/versionhistory.rst
@@ -3,7 +3,7 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
-**UNRELEASED**
+**3.0.0rc2**
 
 - Changed the asyncio ``CancelScope`` to raise a ``RuntimeError`` if cancel scopes are being exited
   in a different order than they were entered